+++ resources/org/corfield/new/template/test/main_test.clj
@@ -1,7 +1,13 @@
 (ns {{top/ns}}.{{main/ns}}-test
-  (:require [clojure.test :refer :all]
+  (:require [clojure.edn :as edn]
+            [clojure.java.io :as io]
+            [clojure.spec.alpha :as s]
+            [clojure.test :refer :all]
+            [org.corfield.new] ; for the Specs
             [{{top/ns}}.{{main/ns}} :refer :all]))
 
-(deftest a-test
-  (testing "FIXME, I fail."
-    (is (= 0 1))))
+(deftest valid-template-test
+  (testing "template.edn is valid."
+    (let [template (edn/read-string (slurp (io/resource "{{top/file}}/{{main/file}}/template.edn")))]
+      (is (s/valid? :org.corfield.new/template template)
+          (s/explain-str :org.corfield.new/template template)))))